requirements.txt
flask==3.1.0
flask-socketio==5.4.1
google-api-python-client==2.154.0
google-generativeai==0.8.3
pytz==2024.2
pandas==2.2.3